Major changes
All files renamed by removing the '2JMP' sub text and changed 'Writer' to 'Importer'.\n Importer instances now store the information in a map located in DAEImporter. The goal is to abstract the importing and exporting. In the future, exporter classes will be added that handle writing the model data to file.